added TravisCI instructions for regression tests
This allows TravisCI to build Machinekit on an amd64 environment and run
the regression tests.

It does this using a Docker container to get a reproducable build
environment for a 64-bit host. Inside the container it uses proot/chroot
to actually do the building and testing.

The current structure is set up this way to have a general structure
that runs on x86_64 hosts which can build for amd64, i386 and ARM
environments.