posemath: correct how "FUZZ" calculation is handled for cartesian vectors
The IS_FUZZ macro decides if a vector is long enough to be considered
nonzero. Previously this was implemented by comparing the magnitude
(i.e. 2-norm) to a minimum length (CART_FUZZ). If the magnitude is
shorter than that length, it's treated as zero length (and the tmag_zero
flag set).

The problem with this approach is that the canon layer uses a different
metric. By canon standards, a vector is zero-length iff all of its
components are less than a "tiny" amount, currently 1e-7. This means
that a vector of <0.8, 0.8, 0.8> * 1e-7 would be "zero" in canon terms,
but since it has a magnitude of ~1.4e-7, it would be nonzero in TP /
posemath calculations.

The solution is to implement a "supremum" calculation (i.e
infinity-norm) in posemath, and use that norm to set the tmag_zero flag.
As long as canon and posemath share the same cutoff value (tiny vs
CART_FUZZ), both calculations will yield the same result:

CANON:

 x > tiny or y > tiny or z > tiny implies non-zero

POSEMATH:

supremum = max(abs(<x,y,z>)), supremum > CART_FUZZ

if tiny = CART_FUZZ, then both are the same calculation.

Signed-off-by: Robert W. Ellenberg <rwe24g@gmail.com>