tests: verify that the exported realtime math functions exist, except for round()
This verifies that the math functions declared in rtapi_math.h are all
actually available at load-time.

... except for round().

Our rtapi_math.h has included a declaration for round() since 2006,
but the LinuxCNC RTAI packages up through RTAI 3.8.1 (which we packaged
for Ubuntu Lucid) did not actually include an implementation of round().

This commit is going into the 2.7 branch, which is the last branch to
support Lucid, so I'm disabling the round() part of the test here.

RTAI 3.9 and 5.0-test1 (packaged for Debian Wheezy and Jessie, and Ubuntu
Precise) do include round(), so the round() part of the test will be
enabled in linuxcnc branches after 2.7.

Signed-off-by: Sebastian Kuzminsky <seb@highlab.com>