tests: add a test of the rtapi mutex operations
this simplistic test increments a counter that is protected by the rtapi
mutex.  If the mutex works properly, the counter is equal to a known
value (number of iterations times number of threads).  If the mutex
doesn't work properly, the counter can have a value other than that
known value.

This test passed 10,000 trials on an x86_64 machine and 10,000 trials on
an ARM machine.  It readily fails when no-op mutex operations are
substituted instead.  This gives us a level of confidence that the mutex
primitives are correct.

Signed-off-by: Jeff Epler <jepler@unpythonic.net>