canon: handle degenerate segments in NCD distance test
In this case, degeneracy refers to the case where the current "end"
point in canon is very close to the newest point. In the NCD
calculation, These two points form the line segment; then, we find the
distance between each "chained" point and this line segment. If the new
point being added happens to coincide with the "end" point, then the
line segment will degenerate to a point. Instead of failing here due to
a divide by zero, we can instead assume that distance to the line ~=
distance to an endpoint.

Signed-off-by: Robert W. Ellenberg <rwe24g@gmail.com>