motion: Remove confusing use of emcmotglb from realtime
These variables are intended to be used in milltask.  However, because
the emcmotglb object file was linked to the modmod realtime component,
there was a separate and unininitialized copy of these values there as
well.

This leads only to confusion, such as with the (now modified)
initialization of emcmot structure fields from VELOCITY and ACCELERATION
values that would always be equal to 0.0.

For the same reason, remove it from the set of public header files.

This confusion also tripped up Moses when he tried to initialize a new
emcmotglb.c field "inhibit_probe_errors" from task and then use the value
in motion.

Signed-off-by: Jeff Epler <jepler@unpythonic.net>