component: panelui -add keyboard to command  ui decoder
uses an INI file to designate commands and/or HAL pins.
allows a handler file to add custom commands in python
buffers keyboard scan codes in a shared memory buffer
current design expects sampler to fill the shared mem.
expected that the sampler wuill be connected to
mesa 7i73 or matrix_kb component for scan codes.

Signed-off-by: Chris Morley <chrisinnanaimo@hotmail.com>