interp: don't return potentially stacked data
The old formulation was always wrong, because it constructed a temporary
string and returned the value returned by its .c_str() method.

Historically in libstdc++, this second string had the same .c_str() as
the other one.  However, in the version of libstdc++ on Fedora 23
(libstdc++-5.1.1-4.fc23.x86_64) this is not the case.

This led to messages such as
	cant resolve path to '??????'
as well as valgrind diagnostics and possibly to the diagnostic "'utf8'
codec can't decode byte".

Signed-off-by: Jeff Epler <jepler@unpythonic.net>