debian: accept the new style of rtai-modules package name
Starting with the Linux 3.16 RTAI 5.0-test1 kernel, we're building for
both the i386 (686-pae) and amd64 architectures, and the rtai-modules
debian package changed name to not include the kernel flavor ("rtai")
or the architecture ("686-pae" or "amd64").

On Wheezy it was named "rtai-modules-3.4.0-9-rtai-686-pae" and was
available just on i386, but on Jessie it is named "rtai-modules-3.16.0-9",
and is available for both i386 and amd64.

This commit changes how debian/configure writes debian/control.  It now
accepts either old-style or new-style names for rtai-modules.

Signed-off-by: Sebastian Kuzminsky <seb@highlab.com>