emcmodule: Fix incorrect memory access by PyArg_ParseTuple and add better checks for string arguments
Jeff Epler note about set_digital_output:
The underlying type of m.index and m.start is unsigned char, but "i"
stores as many bytes as there are in a platform integer.  This could
lead to a crash as reported by micges, and probably also led to
incorrect digital outputs too.
Unfortunately, the compiler does not type-check Python argument
specifiers, so this mistake passed silently.

Signed-off-by: Michael Geszkiewicz <micges@wp.pl>