ngcgui: new variable for ct of remaining features
global: #<_remaining_features:> indicates the number of features
remaining after the current feature.  A value of zero indicates that
the current feature is the last feature.

A user subroutine (or preamble or postamble) can test
value to display count of remaining features and/or to test
if the current feature is the last feature in a sequence.

Also: include debug prints in some examples

Credit to RickG for suggestion on the LinuxCNC forum

Signed-off-by: Dewey Garrett <dgarrett@panix.com>