keystick: fix signal handler a second time
With key repeat settings
    xset r rate 200 20
this one seems to behave correctly, and as keystick did with the 2009
fix reverted, but without worries about the range of activities
performed in the SIGALRM signal handler.

With the (default)? key repeat settings
    xset r rate 500 20
each key erroneously repeats once (so e.g., holding F1 while in estop
toggles to estop reset and back to estop; pressing right arrow in
incremental jog mode does two incremental jogs), but this is the same as
with the patch reverted.  The code is inherently tied to working with
specific key repeat values.

Signed-off-by: Jeff Epler <jepler@unpythonic.net>