hostmot2: sserial: work around sticky error reporting
In sserial firmware <=43, the error bits in the status
register stick instead of being reset when a 'doit' message
is sent.  So rely on the data in the read register as the
first indication of an error.  Only print information about
errors denoted in the status register once the error limit
is reached.

Signed-off-by: Jeff Epler <jepler@unpythonic.net>