hostmot2: sserial: actually check if read function added
Though I never deciphered it, and seb_kuzminsky observed
that the "comm_err_flag" involved in the condition was
only ever set to zero, the heuristic looks like it was
intended to check if all the configured channels on an
instance got an error.  This heuristic didn't work; even if
properly implemented, it would fire for anybody who has a
single sserial device attached.

Instead, explicitly set a flag to denote that the
process_tram_read function was ever reached, and show the
message exactly in the case that it was never reached.

In the process of doing this, several variables became unused.
Remove them.

Signed-off-by: Jeff Epler <jepler@unpythonic.net>