xhc-hb04.tcl: accomodate prior connect of this pin:
motion.spindle-speed-out-rps-abs

pncconf & stepconf connect this output pin to a signal even
when it is not used (e.g., no connection to an input pin).

So users attempting to use xhc-hb04.tcl are confused by error
message about an existing connection since the signal is not used.

This commit finds that signal (if it exists) and uses it instead
of trying/failing/reporting the attempt to create a new signal.

Signed-off-by: Dewey Garrett <dgarrett@panix.com>