motion: remove heuristic delay warning
This check was added at a time when we did not log missed deadline
warnings in the underlying RTOS.  Since all RTOSes now detect a missed
cycle, this heuristic warning is not needed.  Additionally, many users
report that the warning appears even on systems that work properly, so
it is actively harmful.

Signed-off-by: Jeff Epler <jepler@unpythonic.net>