axis.py,glcanon.py detect historical lathe configs
and make work with updated trivkins (sequential joint
assignment to axis letters, formerly gentrivkins)

Prior to joints_axes, many configs used a lathe config
with a dummy joint0/Yaxis.  This was done, I believe,
largely for convenience of hal setups (core_sim.hal and
friends).  This type of config is also used in lathe
configs in the by_machine and by_interface directories.

This commit removes an earlier hack for such lathes and
adds a new detection to handle:
  is_lathe && coordinates=="XZ" && joints==3

This allows the update_ini script provide a working config
from the historical lathe configs and retention of
halfiles that assume dummy joint0/Yaxis for lathes, namely:
   LIB:core_sim.hal
   LIB:simulated_home.hal
   LIB:lathe.hal

Signed-off-by: Dewey Garrett <dgarrett@panix.com>