initraj.cc restore [TRAJ]HOME use (hexapod fail)JA
The hexapod sim config (configs/sim/axis/vismach/hexapod-sim/hexapod.ini)
has been broken in joints_axesNN branches since joints_axes8 (at least)
while it works in recent master (SHA 669f5fe).

After starting the sim, a switch to world mode (before or after homing)
results in all 'nan's for the world coordinates.  The same behavior
can be shown in master by removing the ini setting:
   [TRAJ]HOME = 0 0 20 0 0 0
e.g,. the sim config fails using default zero values for [TRAJ]HOME.
In testing, the sim config succeeds for many values for the z coordinate,
like:
     #nan HOME = 0 0 0 0 0 0
     #ok  HOME = 0 0 10 0 0 0
     #ok  HOME = 0 0 1 0 0 0
     #ok  HOME = 0 0 .1 0 0 0
     #ok  HOME = 0 0 .0001 0 0 0
     #nan HOME = 0 0 .00001 0 0 0
     #ok  HOME = 0 0 -1 0 0 0
     #ok  HOME = 0 0 40 0 0 0
     #ok  HOME = 0 0 100 0 0 0
     #ok  HOME = 0 0 1.23 0 0 0

The problem is due to a failure of a matrix inverse computation when the z
coordinate==0 in the kinematicsForward() function of genhexkins.c (note it
works for z=0.0001 but not 0.00001).

For JA, the code to read and use inifile item [TRAJ]HOME was omitted years ago
so this commit restores it making it possible to use non-zero values for the
[TRAJ]HOME z coordinate.

This problem may be solvable in genhexkins.c --only the omission of
inifile item [TRAJ]HOME is addressed with this commit.

The problem is likely only reproducible when using kinematics modules that use
iterative algorithms in kinematicsFowrward().  The genhexkins.c is the only
current kinematics module that does this.

Also:  taskintf.cc: removed superseded use of: EmcPose home

Note: The vismach model view is not correct for the hexapod sim in JA, it looks
      ok in master.

Signed-off-by: Dewey Garrett <dgarrett@panix.com>