GladeVCP - New Widget SpeedControl - Adjustment for Touch Screen
The SpeedControl Widget serves as a slider with button to increment or
decrease the value and a progress bar showing the value with or without
units.
It is designed to be used with touch screens

It is needed to change the sliders used with gmoccapy.
That changes will be pushed separately.


Signed-off-by: Norbert Schechner <nieson@web.de>