motion/command.c EMCMOT_TELEOP coordinating=0   JA
This appears to be a long-standing omission that had no consequences
prior to joints_axes.  In joints_axes, properly setting
   emcmotDebug->coordinating = 0
when the mode is changed to teleop eliminates discontinuous
positions if free mode jogging has occurred.  Such discontinuities
can lead to a joint following error.

This commit remove workarounds in axis.py that forced joint mode
before mdi, task_run, program_open operations because of
the potential for position discontinuities

Signed-off-by: Dewey Garrett <dgarrett@panix.com>