Task: allow EMC_TASK_PLAN_SYNCH in all modes and states
Before this commit, Task used to only allow the UI to sync the Interpreter
if the machine was in the On state, and Task was in the Manual mode.

This commit makes Task also allow the UI to sync the Interpreter in
Estop, Estop-Reset, and Off states, and in the On state with Task in
the MDI mode.

On/Auto is still disallowed because I don't need it for now.

Signed-off-by: Sebastian Kuzminsky <seb@highlab.com>