hostmot2: sserial: avoid glitches when starting to run
With a 7i73, I noticed that in the first cycle after .run
transitioned from FALSE to TRUE, the input pins all gliched
to FALSE.  This is because in the .write function in cycle #N,
inst->state changes from 0x11 to 0x1, but no read instruction
is issued yet.  In cycle #N+1, this causes stale or zeroed data
to be used, making a physical pin that is driven high to glitch
FALSE.

By introducing the new 0x4/0x5 states into the state machine,
the first .read invocation after transitioning to run keeps pin
values held, and the glitch is suppressed.

Signed-off-by: Jeff Epler <jepler@unpythonic.net>