configs/sim/axis_mm.ini: move the explicit thread creation
This shows how motion can be prevented from creating a thread, and
rather create the thread explicitly in hal.

Note the -[EMCMOT]SERVO_PERIOD parameter might be referenced by other
HAL configs (eg core_sim.hal) so must be set to 0 as long as this is the
case, not deleted.