Added callLevel to EMC_TASK_STAT class, to fix hal_glib file-loaded bug.
callLevel will be non-zero if executing a subroutine.  When subroutines are
executing, EMC_TASK_STAT::file gets set to the file name of the subroutine
(remap as well I think).
Exporting the call level in the task status allows us to ignore file name
changes in hal_glib and avoid sending bogus signals to things
such as hal_gremlin and hal_sourceview.

Signed-off-by: Moses McKnight <moses@texband.net>