tests/motion-logger/basic: allow comments in expected files
This augments the motion-logger/basic test to allow comments in the
'expected' files.  The 'expected' files are preprocessed to remove
comment lines and empty lines before being compared to the actual output.

The 'expected.builtin-startup' file is renamed with a '.in' extension to
trigger this kind of preprocessing.  No actual test content was changed
with this commit.

Signed-off-by: Sebastian Kuzminsky <seb@highlab.com>