stepconf: Always write postgui_call_list.hal
I created a sim configuration with all defaults, but it didn't run because
postgui_call_list.hal didn't exist.  It appears this would have affected any
configuration that didn't set the pyvcp option.

It doesn't appear critical for the condition to match the one under which
lines that *do things* are written to pyvcp_options.hal, because that file
is created unconditionally just below the context of this change.

Signed-off-by: Jeff Epler <jepler@unpythonic.net>