uspace: fix locking of shared memory segments
I misread the shmctl manpage, and tried to use shmctl IPC_SET
with the SHM_LOCK bit set in shm_perm.mode.  Instead,
the right way to do this is shmctl SHM_LOCK, and then check
for the SHM_LOCK bit.

Only do this in rtapi_app when rtapi_is_realtime(), because
it doesn't matter if a segment can't be locked by a non-
realtime process.  This also prevents needing to raise the
locked memory limit, because rtapi_app has used its root
powers to raise RLIMIT_MEMLOCK.

Signed-off-by: Jeff Epler <jepler@unpythonic.net>