latency-histogram: set pixel size of window explicitly
When testing, I noticed that if X detected an unusually low screen
DPI, the size of the graph area in latency-histogram would be small
relative to the rest of the window.

This is because the default size of a blt barchart is in inches!  By
specifying pixel sizes, the window will get the same layout no
matter the screen's detected DPI.  The numbers chosen are close to
the original values if DPI=96, a typical value assumed for normal
computer monitors (for instance, a 23" full HD monitor or a 15",
720p monitor)

Signed-off-by: Jeff Epler <jepler@unpythonic.net>