stepconf: take nicokid's version of these files
I made a mistake when rebasing nicokid's branch for issue #33, and
lost the changes for these files.  I simply took the ones from that
branch with the command
    git checkout gh-nicokid/master -- ./axis[uv].glade
while in the src/emc/usr_int/stepconf directory.  at that time,
the branch pointed at
    6747746 Merge branch 'master' of https://github.com/nicokid/linuxcnc

This time I carefully looked over all the differences shown by
    git diff 6747746 44c5fb9 -- src/emc/usr_intf/stepconf/
and they are changes deliberately introduced by me to fix or work around
stepconf bugs, plus the differences that this commit pulls in.

Signed-off-by: Jeff Epler <jepler@unpythonic.net>