Task: fix a recent "surprise motion on abort" bug
Before aborting the Interpreter, sync the Interpreter's idea of
where it is/should be, from Motion's idea of where it currently is.

This has the side effect of clearing the Naive Cam Detector.

Aborting the Interpreter causes (among other things) a Canon call to
reset the G5x offset, which involves a flush of the NCD to Motion.
Without the sync here, any waiting NCD segments will get flushed and
cause unwanted motion during Abort.

Signed-off-by: Sebastian Kuzminsky <seb@highlab.com>