axis: remove old button theme hack
prior to Tk 8.4 (?), radiobutton and checkbutton widgets had a
motif-style appearance that looked unusual compared to the
typical style of radiobuttons.  I included a hack for placing
images over the button indicators.  But in current versions of
Tk, these widgets look about like one expects, so the hack can be
dropped.

Signed-off-by: Jeff Epler <jepler@unpythonic.net>