axis: fix listbox width
Before this commit, the requested width of the history listbox
was actually just the width of the scrollbar packed inside it
(.sby).  Setting "pack propagate" to 0 fixes this, but then the
overall width of the mdi tab grows, because the -width 40 of the
listbox is wider than the -width 40 of the active codes display(!)
because they have different fonts.

Signed-off-by: Jeff Epler <jepler@unpythonic.net>