Provide specific warning of Mac and DOS line endings comp/instcomp
and exit.

When comp or icomp files are read in, the `\n` line endings are used
to split the file.

If `\r\n` DOS style or `r` Mac style endings are present, it plays
havoc and a non specific failure results.

'Borrowed with pride' from jepler musings on same in
Emc-developers Digest, Vol 123, Issue 5

Signed-off-by: Mick <arceye@mgware.co.uk>