emctaskmain.cc handle on_soft_limit #96         JA
Make 'on_soft_limit' available to task and avoid abort looping if
'on_soft_limit'.  A switch to joint mode may be needed for recovery
(as noted in error messages).

Note:

IF
      1) KINEMATICS_IDENTITY
AND
      2) System misconfigured such that joint limits are more
         restrictive than axis limits
AND
      3) User gui provides no means to switch to joint mode
THEN
      System will probably need to be restarted and configured
      correctly when a misconfigured soft limit is encountered.
      (Expert users may use halui mode pins and halui jog pins
       and/or joint-mode wheel jogging pins to recover).

Note: There may be undiscovered side effects if there are
errors in addition to the 'on_soft_limit' error

Signed-off-by: Dewey Garrett <dgarrett@panix.com>