Allow specifying of device on wj200_vfd init
This commit adds the --device option to the wj200_vfd HAL component.
It allows the user to select which device to use for Modbus communication
to the VFD. For example, to use a USB to RS485 converter on /dev/ttyUSB0
you would run this command:

    loadusr -W wj200_vfd --device /dev/ttyUSB0

The default is /dev/ttyS0 if --device is not specified.

Signed-off-by: James Waples <jamwaffles@gmail.com>