gmoccapy_JA_master - solved the MDI History bug
hal_mdi_history used the hal_action_base.is_all_homed
to decide if it should be sensitive or not.
As this function does check for axis not for Joints it will not work
on non identity kinematics and for machines with "non used axis"

If you define a machine with COORDINATES = XYZAC
and JOINTS =  5 what should be possible, axis C will never be homed,
as it is axis 6 but will use Joint 5

Just defining 6 Joints as it has be done for identity machines,
will bring additional Joint Jog buttons, causing confusion for
the users.

Signed-off-by: Norbert Schechner <nieson@web.de>