uspace: avoid use of message queue in main thread
Several tests (including hm2-idrom and module-loading) rely on
capturing output of a halrun session and grepping inside it; this in
turn creates an assumption about how output from rtapi_app (startup
and loadrt-time) and halcmd is interleaved.

With the introduction of the message queue, this assumption was
broken.  Attempt to put back the assumption by bypassing the message
queue when the message is being printed from the "main thread",
i.e., the one that creates the App() object and responds to
load/unload/etc commands.

Comparing pthread_t objects to 0 or to another pthread_t with == is
not portable, but on common platforms it's either an integral type
or a pointer (verified by compile on linux and glance at header on
freebsd, plus
http://stackoverflow.com/questions/30867779/correct-pthread-t-initialization-and-handling
which has to cite something weird like dgux for an exception)

Before this change, tests/module-loading/pid/num_chan=16 failed with
a frequency of between 1-in-100 and 1-in-500 runs on my odroid u3.
After this change, it succeeds for 3000+ iterations without error,
the same as before merging the uspace-plus branch.

Signed-off-by: Jeff Epler <jepler@unpythonic.net>