limit3.comp: add 'in-limit' output pin
The 'in-limit' pin is false when the output is exactly
tracking the input, and true when position, velocity,
or accel limit causes the output to differ from the
input.

Also replaced test limit3.1 (which was a duplicate
of limit3.0) with a more complete test that verifies
accel and velocity limitation and the 'in-limit' pin.

Signed-off-by: John Kasunich <jmkasunich@fastmail.fm>