halui: correctly report "mode.is_joint"
Before this commit, halui's mode.is_joint pin would go True whenever
LinuxCNC was not in Teleop mode.  So it would incorrectly go True when
running a G-code program, since in that state LinuxCNC is in Coordinated
mode instead of Teleop.

This commit fixes halui to only set mode.is_joint True when in Free mode.

Signed-off-by: Sebastian Kuzminsky <seb@highlab.com>