update_ini: Avoid making syntactically incorrect "setp" statements
In cases where a single pin is replaced by two pins, for example the mpg
jogging pins which now have separate joint and axis versions, the script
would put both new pins in one "setp" statement, leading to a HAL parsing
error.
I tried to avoid creating the bad statements in the first place, but it is
harder than it seems, so now the code simply repairs them.

Signed-off-by: andypugh <andy@bodgesoc.org>