Move hal_bb_gpio manpage from docs/src/ to docs/man/
Eventually all man pages will be asciidoc in docs/src/ and
the troff pages in docs/man/ will be auto-generated.  But
that may take a while, and in the meantime troff pages in
docs/src/ aren't usable.  When the move comes in the future
this page can move with all the rest.

Signed-off-by: John Kasunich <jmkasunich@fastmail.fm>