Provide workaround to keep xhc-hb04 pendants working
This workaround is a temporary measure to keep the xhc-hb04 pendant working.
xhc-hb04.tcl uses commandline halcmd outputs against tests which are fixed
in expectation of a particular return. (ie from linuxcnc circa 2014)

Changes to halcmd print routines broke xhc-hb04.

This will tell the tcl code if a pin exists, removing the need for the
previous convoluted parsing of the stdout output.

Signed-off-by: Mick <arceye@mgware.co.uk>