rs274ngc-startup test: wait for Task to start up
Before this patch, sluggish startup of Task & Interp could cause false
failures of this test.

Wait for Task to finish initializing the Stat buffer, and for Interp to
be Idle, before asserting that the RS274NGC_STARTUP_CODE ran successfully.

Verified by (temporarily) changing the tests's startup code to "G4 P15",
the wait code timed out and threw an exception.

Signed-off-by: Sebastian Kuzminsky <seb@highlab.com>