tp: new formula for spindle position tracking
The original "error velocity" formula was correct if the initial and
final velocity was zero. However, during spindle position tracking, the
target velocity is non-zero (or it wouldn't be tracking!). This lead to
a slight over-correction, and steady-state jitter even with a perfect encoder signal.

The new formula accounts for the ideal target velocity, and should
therefore over-correct less. This fixes the jitter (in simulation with a
perfect encoder, at least).

Signed-off-by: Robert W. Ellenberg <rwe24g@gmail.com>