glcanon.py: Allow easy customisation of the DRO number format
Requests to add more (or fewer) significant digits to the DRO for specific
machines or to force a + sign are not uncommon. This commit saves users
from the horror of editing the Python code themselves.
A basic validity check is made, but otherwise no restrictions are made:
hexadecimal, binary, exponent notation are all possible.


Signed-off-by: andypugh <andy@bodgesoc.org>