Add `mdi-while-queuebuster-waitflag` regression test
This test tries to hit a race condition where an MDI command is
executed when a queue buster is waiting to be restarted after
INTERP_EXEC_FINISH.

I don't know how to hit this condition reliably, and this test may not
hit the race on other systems.  On my Intel Core 2 laptop in sim mode,
it hits the race by the third iteration about 90% of the time.

Fix is to follow.

Signed-off-by: John Morris <john@zultron.com>