canon: Enable smarter blending in spindle-synchronized motion.
Since Canon knows the spindle speed for a given segment, it's trival to
calculate the nominal feed for a synchronized motion segment (assuming
ideal spindle behavior). This nominal velocity is passed to motion so
that blend arcs can be sized correctly.

Also, report a canon error if the planner can't meet the required feed
for spindle synchronization.

Signed-off-by: Robert W. Ellenberg <rwe24g@gmail.com>