Build system:  remove `ARCHITECTURE`
This variable was apparently intended to represent the host
architecture, but since it used the `binutils` `arch` command, it was
unaware of cross-build environments.

This patch removes all traces of it, and uses `$host_cpu` in the one
place it was actually used.

Reverts part/all of 5ee8416c, b0e17c03, ae9f0273