docs: remove unused (((indexterm))) that break html
In asciidoc, (((IndexTerm))) is used in the docbook backend, but not
in the html or pdf backends that we use.  A bug in asciidoc results in
messed-up html when (((indexterm))) is used, empty <title>.

Since we don't use docbook, (((indexterm))) does nothing useful for us,
so it seems simplest to just remove it.

Thanks to tseufl for figuring this one out.

Signed-off-by: Sebastian Kuzminsky <seb@highlab.com>