docs: work around a pile of asciidoc wrinkles
In asciidoc:

    * A [[BlockId]] normally applies to the immediately following part
      of the document.

    * An (((IndexTerm))) is used in the docbook backend, but not in the
      html or pdf backends that we use.

Before this commit, the leading [[cha:basic-hal-reference]] applied to
the (((Basic HAL Reference))).

However, a bug or wrinkle in asciidoc made the (((indexterm))) mess up
the html backend, and the cleanest fix seems to be to just remove it
(since it doesn't do anything useful for us).

However, another wrinkle/bug in asciidoc means that a [[BlockId]] before
the document title gets applied not to the title header but to the part
of the document *following* the title.

In this particular document, the document title is followed immediately
by a second-level header with its own [[BlockId]], so the first BlockID
gets lost.  That's not ok, because other files in our docs link to both
the top-level BlockID and the second-level BlockID.

So this commit adds a decoy document element for the top-level BlockID
to apply to.  I am not proud.

Signed-off-by: Sebastian Kuzminsky <seb@highlab.com>