Add `CPPFLAGS` and complete `LDFLAGS` support
`CPPFLAGS` are flags to the C preprocessor.  The build system doesn't
currently make use of this, instead sticking everything into `CFLAGS`.

Add `CPPFLAGS` everywhere, and add `LDFLAGS` missing in several places.

`CPPFLAGS` is important for cross-compiling, because autoconf makes
correct use of it in macros, and the `gcc --setroot` flag isn't passed
through in `CFLAGS`.

Also, pass `LDFLAGS` and `CPPFLAGS` to build when set in the
`./configure` environment.  This may be wrong, but pass those into
build through `Makefile.inc`.