tests: add an abort-vs-feed-rate test (skipped)
This test verifies that the "active g-codes" F-word is correct after
a synch (eg, after an Abort or after a switch into and back out of
Manual mode).  It replicates github Issue #180, reported by @fixerslo.

The test is marked "skip" for now, i'll remove the skip marker after I
fix it in the next few commits.

Signed-off-by: Sebastian Kuzminsky <seb@highlab.com>