Add G92.2,3 to Active Gcodes, state tags, etc.
This puts G92.1, G92.2 and G92.3 in group 16 in order to show in Active GCodes
whether G92 is currently applied.  The readout only ever shows G92.2 and G92.3,
since G92.1 is a special case of G92.2.

Wheter G92 is applied (aka variable 5210) is added to the state tags, and
restored on abort.

Previously, aborting a program that used G92 offsets would give you unapplied
G92, contrary to expectations.

Signed-off-by: Chris Radek <chris@timeguy.com>