Implement 9d84d0d -- deprecate singleton in instantiated components
The option was originally supported for backward compatibility,
however it really contradicts the basic concept of instantiation.

If a component cannot ever have more than one instance, it should
be written as a standard .comp

Arose initially from an Issue where checking for other instances
of a component marked as a singleton, without proper mutex lock
was causing problems
On reflection the whole concept within instcomps was removed.

Whilst editing instcomp, corrected an error that had crept in during
changes to document generation, whereby the pin names were displayed
in the docs in C style rather than the HAL names

Signed-off-by: Mick <arceye@mgware.co.uk>