tp: Re-order checks for parabolic blends
Parabolic blends require reduced max acceleration, so that the
worst-case blend does not exceed the limits. If a parabolic blend can be
"upgraded" to a tangent / arc blend, then we don't need this
restriction. Previously, this check occured too early in the process of
creating a new segment, which meant some segments ended up with reduced
acceleration unnecessarily.

Signed-off-by: Robert W. Ellenberg <rwe24g@gmail.com>