tp: Remove discontinuous spindle positions in rigid tapping internals.
Rigid tapping used to estimate spindle position both from the measured
spindle revs, and the commanded direction. This approach is numerically
dangerous, however, because the spindle position itself could be large.
Therefore, flipping the sign on the position would cause an apparent
jump. The previous approach to rigid tapping accounted for this.
However, this new approach should be cleaner and simpler.

Signed-off-by: Robert W. Ellenberg <rwe24g@gmail.com>