changed the __init__ function making the renderers all editable all of the time.
Added the onTreeNavigateKeyPress function which handles key presses in all three tree view widgets. The tab and right arrow keys will accept the current value in the cell and progress to the next visible cell to the right. If there is no cell to the right then it moves to the first visible cell in the row. The left arrow key is the same except it moves left. The enter key and the down arrow key accepts the current entry in the cell and moves to the next row same column. If on the last row then it will move to the top row. Up arrow key is the same but moves up instead.

Signed-off-by: Jim Craig <jimcraig5615@windstream.net>