stdglue.py:  turn `change_epilog()` into generator function
The functionality to return `INTERP_EXECUTE_FINISH` in a python
remap function has always existed, but was broken until PR #135.

The fix revealed a bug in the standard remap `change_epilog()`
function, which was not written as a python generator function.

Thanks to @dhwiezel for discovering and reporting this bug.  Fixes
#217.

Signed-off-by: John Morris john@zultron.com