Amend halcmd print routines to expand name length to 40 chars
This modifies previous commit down from 49 chars, because although
a new length of HAL_MAX_NAME_LEN of 127 is introduced in hal.h
it is not generally used.

This means for example,any pin name of over 41 chars will not
be able to be linked to a signal, rendering the 127 name limit
of little use.

Signed-off-by: Mick <arceye@mgware.co.uk>