Unit test for maxvel violations in spindle-synched moves
In a spindle-synched G33 move, it is possible to specify a feed per
revolution and spindle speed that requires linear motion exceeding
axis velocity limits.  This test sets up this kind of condition and
measures actual pitch during the motion, failing if pitch is not
within an epsilon value.

See LCNC #167 for more discussion.

https://github.com/LinuxCNC/linuxcnc/issues/167

Signed-off-by: John Morris <john@zultron.com>