canon: don't restore spindle speed after limiting it for spindle-synched motion.
Feed / speed synch is always disabled after a synced segment (even
if it's then re-enabled for the next one). Unfortunately, this means
that limiting spindle RPM, and then restoring it, has the effect of
inserting two spindle speed commands between each segment of synched
motion. This completely disrupts blending, and affects the acceleration
used during the motion (potentially affecting threading performance).

A safer and easier approach is to pop up a nuisance message to the user
informing them that the spindle speed will be limited, and then just
leave it at the limited speed. Generally, the user will set a new
spindle speed for the next operation anyway.

Signed-off-by: Robert W. Ellenberg <rwe24g@gmail.com>