Rename new C components and fix runtests
Components renamed v2 without preceeding dash to be compatible
with all instcomp generated ones.
Any config can be altered for v2 components simply by adding v2
to the existing names.

pin_prefix set to same as the component name
Not sure how it causes 'less surprises' to have the function prefix
for a v2 component exactly the same as a v1 component and not the
same name as you instantiated.
You can't tell which one you are looking at in a print.
Original author not around to ask!

Fix up runtests to use changes as above

Signed-off-by: Mick <arceye@mgware.co.uk>