Added HAL-OUT pin for ignore-limits-active
Added HAL (OUT) pin that is set TRUE when the "Ignore Limits" checkbox is active, and FALSE all other times. Allows, e.g., for a PLC to control release of limit switches, and be notified by GMOCCAPY/LinuxCNC that the operator is overriding the limits (without adding a physical override button).