travis: manually uninstal gpl3 readline
In the past, the apt resolver automatically uninstalled the gpl3
readline (libreadline-dev, libreadline6-dev) as needed to satisfy
the dependencies of linuxcnc-build-deps.  At some point around
2017-02-10, this changed on travis-ci.org and instead it preferred
to remove linuxcnc-build-deps and leave a bunch of deps (such as
libudev-dev, the first one configure hit) in a non-usable state.
Work around this by manually removing the packages that seem to be
causing the problems.

Signed-off-by: Jeff Epler <jepler@unpythonic.net>