uspace: allow calculated parameter array sizes
In the original formulation, a use like
  RTAPI_MP_ARRAY_INT(arr, 3*FOO, "array of three times foo items");
gave a weird error on uspace only when arr= was actually
specified on the loadrt commandline.  This is because the 'num'
value was communited to rtapi_app as a string rather than an
integer, so the string "3*FOO" was received where a string like
"24" was expected, resulting in an error.

Happily we have good test coverage for comp's use of
RTAPI_MP_ARRAY via tests/module_loading/or2* so even though this
was encountered and reported in the non-runtested hal_ppmc driver,
we can have confidence about the solution.

Close #242.

Signed-off-by: Jeff Epler <jepler@unpythonic.net>