docs: fix gmoccapy asciidoc markup
In Asciidoc a string in square brackets is an 'attribute list' that gets
passed to the code that deals with the following block element.

This commit changes the attribute lists of a couple of 'paragraph'
block elements.  Paragraph blocks expect their attribute list to identify
which paragraph style to use.

'Note' is not a known paragraph style, so asciidoc complains:

    asciidoc: WARNING: gmoccapy.txt: line 680: missing style: [paradef-default]: Note

It then renders the following paragraph in the default style.

This commit changes the attribute from "Note" to "NOTE" (in all caps),
so that asciidoc uses the "NOTE admonishion" paragraph style for those
paragraphs.

Signed-off-by: Sebastian Kuzminsky <seb@highlab.com>