axis: avoid unbounded memory growth in text widgets on stretch
I reproduced a memory leak problem in AXIS most readily when
switching to the text DRO tab (F4).  This situation may trigger when
the window is withdraw, minimized, on a different desktop, or fully
hidden by a screensaver, because under these conditions the DRO tab
is updated regularly(!).

In valgrind, the leak looks like
    1,130,496 bytes in 69 blocks are possibly lost in loss record 6,257 of 6,257
       at 0x4C2BBAF: malloc (vg_replace_malloc.c:299)
       by 0x83A6971: GetBlocks (tclThreadAlloc.c:1044)
       by 0x83A6971: TclpAlloc (tclThreadAlloc.c:358)
       by 0x82C9607: Tcl_Alloc (tclCkalloc.c:1059)
       by 0x7F648FF: Tk_QueueWindowEvent (tkEvent.c:1641)
       by 0x7F80B08: TkSendVirtualEvent (tkUtil.c:1193)
       by 0x7FE2838: TkTextInvalidateLineMetrics (tkTextDisp.c:3377)
       by 0x7FDC750: TkBTreeDeleteIndexRange (tkTextBTree.c:1557)
       by 0x7FD8B5F: DeleteIndexRange (tkText.c:3273)
       by 0x7FD2F8D: TextWidgetObjCmd (tkText.c:1113)
       by 0x82C06D6: TclNRRunCallbacks (tclBasic.c:4390)
       by 0x79A8EEF: Tkapp_Call (_tkinter.c:1653)
       by 0x2033C9: call_function (ceval.c:4352)
       by 0x2033C9: PyEval_EvalFrameEx (ceval.c:2989)

Due to a bug in Tk 8.6.6, text widgets on stretch come configured
with "-undo 1 -maxundo 0", leading to unbounded memory use as text
widget contents are updated.  Set -undo to 0 so that the undo
facility is explicitly disabled.

I bisected this as a bug in the tk git repository:
    git bisect start core_8_6_6 core_8_6_2
    git bisect run sh -c 'cd unix; make clean; ./configure ||
        exit 125; make -j5 || exit 125; make shell SCRIPT=tx.tcl'

tx.tcl is a script which determines whether the .text -undo
option is zero (OK) or nonzero (failing):
    text .t
    exit [.t cget -undo]

This is a bug in Tk that was corrected after the 8.6.6 release:
    2f6662d9dd6f78bacd667a48e6aa509ca342369a is the first bad commit
    commit 2f6662d9dd6f78bacd667a48e6aa509ca342369a
    Author: fvogel <fvogelnew1@free.fr>
    Date:   Wed Jun 8 15:27:05 2016 +0000

        Fixed [6976e4b3b2] - -undo options not correct for peers

The fixing commit in tk.git is apparently
    commit b65753c283f6edb19d7f4269c83008eed7f43bdc
    Author: pspjuth <peter.spjuth@gmail.com>
    Date:   Tue Aug 30 19:46:35 2016 +0000

        Made text widget's default undo false again.

This problem potentially affects other Tk-based UIs, but they don't
manipulate the contents of a Text widget as heavily as AXIS does for
its text DRO tab (tklinuxcnc uses Label widgets for this purpose
instead)

The text "-undo" configuration option exists in all Tk 8.4 and newer
versions, so there's no need for the setting to be conditional.

Signed-off-by: Jeff Epler <jepler@unpythonic.net>