Fix time and timev2 instcomps
A complete line had been deleted from the function code,
with the effect that they did not update.

Upgraded to use fa_current_period() so that updates are more accurate.
Also catered for restart, which was never in original
by adding a zeroing pin.

Timings now can either be cumulative, with stops and starts
or reset to zero whilst timing or after a stop.

Signed-off-by: Mick <arceye@mgware.co.uk>