gladevcp -improve background color with themes
This widget use to set the background directly to white.
In some themes the text color would be white so couldn't be seen.
Now it follows the theme suggested background colors.

I took the background color away from the liststore since it caused
an error message when blank.
As there was no helper function, I doubt this option was exercised.
One can still change the background directly with python.
It does require more code and more knowledge to do.

I could avoid the error message by filling in the liststore color
with the system/theme color when realized, but it seems alot of fooling
around for little - the theme background is probably the way to go always..
If people complain I will look into it again.

Signed-off-by: Chris Morley <chrisinnanaimo@hotmail.com>