Fix current ini file detection by gremlin_view
Recently used above in project and found that it failed.
pgrep was failing to detect the inifile name from the running linuxcncsvr.

An errant sed operation had changed it to 'machinekitsvr', but even
when this was corrected, pgrep still errored thus

```
Traceback (most recent call last):
  File "/usr/src/machinekit/bin/gremlin_view", line 3, in <module>
    gremlin_view.standalone_gremlin_view()
  File "/usr/src/machinekit/lib/python/gremlin_view.py", line 597, in standalone_gremlin_view
    ,yoffset=yoffset
  File "/usr/src/machinekit/lib/python/gremlin_view.py", line 157, in __init__
    if ini_setup():
  File "/usr/src/machinekit/lib/python/gremlin_view.py", line 108, in ini_setup
    ini_filename = get_linuxcnc_ini_file()
  File "/usr/src/machinekit/lib/python/gremlin_view.py", line 135, in get_linuxcnc_ini_file
    ans = p.split()[p.split().index('-ini')+1]
ValueError: '-ini' is not in list
```

Amended linuxcnc script to export the inifile name to the environment as INIFILE

Amended gremlin_view to search env for INIFILE and it works

Signed-off-by: Mick <arceye@mgware.co.uk>